requirements_affinity.txt
# Affinity Rescoring - Additional Dependencies
# These are on top of the base boltz requirements in pyproject.toml

# Core scientific
numpy>=1.24.0
torch>=2.0.0

# Structure parsing
biopython>=1.81
gemmi>=0.6.0

# Chemistry
rdkit>=2023.03.1

# CLI & config
click>=8.1.0
pyyaml>=6.0
pydantic>=2.0.0

# Data & export
polars>=0.19.0
openpyxl>=3.0.0     # Excel export

# Progress & logging
tqdm>=4.65.0

# Testing (optional dev dependency)
# pytest>=7.0.0
# pytest-cov>=4.0.0
